eventsource/eventsource-readyState.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>ServerSentEvent Test: EventSource - readyState</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  test( function () {
    var source = new EventSource("resources/message.py");
    assert_equals(source.readyState, source.CONNECTING, "the readyState must be set to CONNECTING");
  }, "Check if EventSource.readyState is CONNECTING when connection has not yet been established");

  var t = async_test("Check if ServerSentEvent.readyState is OPEN when the open event was dispatched");

  var source = new EventSource("resources/message.py");
  source.onopen = t.step_func(function (event) {
    assert_equals(source.readyState, source.OPEN, "the readyState must be set to OPEN");
    t.done(); 
  });
  
  test( function () {
    var source = new EventSource("resources/message.py");
    source.close();
    assert_equals(source.readyState, source.CLOSED, "the readyState must be set to CLOSED");
  }, "Check if EventSource.readyState is CLOSED when the close method was invoked");
</script>
